+++ docutron/src/components/Page.js
@@ -7,28 +7,8 @@ import { Link } from 'react-router-dom';
 /**
  * Internal Dependencies
  */
+import MarkdownContent from './MarkdownContent';
 import { getNextStory, getPreviousStory } from '../config';
-import markdown from '../markdown';
-import Tabs from './Tabs';
-
-function MarkdownContent( { content } ) {
-	const blocks = markdown( content );
-
-	return (
-		<div>
-			{ blocks.map( ( block, index ) => {
-				if ( block.type === 'raw' ) {
-					return <div key={ index } dangerouslySetInnerHTML={ { __html: block.content } } />;
-				}
-				if ( block.type === 'codetabs' ) {
-					return <Tabs key={ index } tabs={ block.tabs } />;
-				}
-
-				return null;
-			} ) }
-		</div>
-	);
-}
 
 class Page extends Component {
 	componentDidMount() {
